Update audioGagger.py
this change made the censoring happen properly where the length of the output audio isn't longer than the initial audio. the censored sound effect is trimmed by the script in order to take up the amount of time that each cuss word lasts. 

next i want to make the tool work on any audio file in a current directory instead of needing to rename the input audio file to "input.mp3"