Update ismr.py
added a dialog box that asks if a video file is being processed, if not then the program pulls up a dialog box for choosing an audio file and runs normally. if a video file is selected then the script will process the transcription and censoring of the audio from the video file and overlay the censored audio over the video file and output a video file with the same name as the input with the word "censored" appended to it.